move entity to the document root
