+++ jnosql-mapping/jnosql-mapping-document/src/main/java/org/eclipse/jnosql/mapping/document/DefaultDocumentTemplate.java
@@ -16,6 +16,7 @@ package org.eclipse.jnosql.mapping.document;
 
 import jakarta.nosql.document.DocumentManager;
 import jakarta.nosql.mapping.Converters;
+import jakarta.nosql.mapping.QueryMapper;
 import jakarta.nosql.mapping.document.DocumentEntityConverter;
 import jakarta.nosql.mapping.document.DocumentEventPersistManager;
 import jakarta.nosql.mapping.document.DocumentWorkflow;
